feat: embed simbriefdatarequest into simbriefdata
